Sortear aleatoriamente uma marmita disponivel no beleaf a partir do(s) tipo(s) de refeicao(oes) e de um preco maximo

Sortear aleatoriamente uma marmita disponivel no beleaf a partir do(s) tipo(s) de refeicao(oes) e de um preco maximo

sortear_marmita(tipo = c("Sopa", "Prato", "Doce"), preco_maximo = 50)

Arguments

tipo
Um vetor de caracteres informando a refeicao: As opcoes disponiveis sao: "Sopa", "Prato", "Doce". O default e c("Sopa", "Prato", "Doce")

preco_maximo
Um inteiro maior que zero

Value

Um vetor de caracteres

Examples

# saida default da funcao
sortear_marmita()

#> Refeicao  sugerida: SEM-CARNE MOÍDA COM LEGUMES 
#> 
#> Ingredientes: Feijão carioca, arroz cateto e trigo integral 
#> 
#> Preco: R$ 20.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/sem-carne-moida-com-legumes

# opcoes de sopa
sortear_marmita(tipo = "Sopa")

#> Refeicao  sugerida: CREME DE ABÓBORA 
#> 
#> Ingredientes: Semente de chia e gengibre 
#> 
#> Preco: R$ 16.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/creme-de-abobora

# opcoes de Doce
sortear_marmita(tipo = "Doce")

#> Refeicao  sugerida: BOLO DE MAÇÃ COM CANELA 
#> 
#> Ingredientes: Açúcar Mascavo e Castanha do Pará 
#> 
#> Preco: R$ 8.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/bolo-de-mac-com-canela

# refeicoes com preco de no maximo R$ 20,00
sortear_marmita(preco_maximo = 20)

#> Refeicao  sugerida: TORTA DE MAÇÃ COM ESPECIARIAS 
#> 
#> Ingredientes: Açúcar Demerara, Aveia e Linhaça 
#> 
#> Preco: R$ 8.9 
#> 
#> Acesse o site: https://www.beleaf.com.br/torta-de-mac-com-especiarias